add class machine

+++ wc_chassis/include/param/speed_v.h
@@ -1,5 +1,6 @@
 #ifndef SPEED_V
 #define SPEED_V
+#include <ros/ros.h>
 
 class Speed_v {
  public:
@@ -11,7 +12,7 @@ class Speed_v {
       double dec;
       double dec_to_zero;
       double full;
-      double remote_level;
+      int remote_level;
 };
 
 #endif
